Enable WarmUp with new re-organization
