+++ src/core/backend.cc
@@ -38,73 +38,6 @@
 
 namespace nvidia { namespace inferenceserver {
 
-namespace {
-
-// Utilities for warmup feature
-TRITONSERVER_Error*
-WarmupResponseAlloc(
-    TRITONSERVER_ResponseAllocator* allocator, const char* tensor_name,
-    size_t byte_size, TRITONSERVER_MemoryType preferred_memory_type,
-    int64_t preferred_memory_type_id, void* userp, void** buffer,
-    void** buffer_userp, TRITONSERVER_MemoryType* actual_memory_type,
-    int64_t* actual_memory_type_id)
-{
-  *buffer = malloc(byte_size);
-  if (*buffer != nullptr) {
-    *actual_memory_type = TRITONSERVER_MEMORY_CPU;
-    *actual_memory_type_id = 0;
-    return nullptr;
-  }
-
-  return TRITONSERVER_ErrorNew(
-      TRITONSERVER_ERROR_INTERNAL,
-      "failed to allocate output buffer for warmup.");
-}
-
-TRITONSERVER_Error*
-WarmupResponseRelease(
-    TRITONSERVER_ResponseAllocator* allocator, void* buffer, void* buffer_userp,
-    size_t byte_size, TRITONSERVER_MemoryType memory_type,
-    int64_t memory_type_id)
-{
-  free(buffer);
-  return nullptr;
-}
-
-ResponseAllocator warmup_allocator = ResponseAllocator(
-    WarmupResponseAlloc, WarmupResponseRelease, nullptr /* start_fn */);
-
-void
-WarmupResponseComplete(
-    TRITONSERVER_InferenceResponse* iresponse, const uint32_t flags,
-    void* userp)
-{
-  if (iresponse != nullptr) {
-    LOG_TRITONSERVER_ERROR(
-        TRITONSERVER_InferenceResponseError(iresponse), "warmup error");
-    // Just delete the response, warmup doesn't check for correctness
-    LOG_TRITONSERVER_ERROR(
-        TRITONSERVER_InferenceResponseDelete(iresponse),
-        "deleting warmup response");
-  }
-}
-
-// FIxME: Move the warmup logic into TritonModelInstance
-//void
-//WarmupRequestComplete(
-//    TRITONSERVER_InferenceRequest* request, const uint32_t flags, void* userp)
-//{
-//  if ((flags & TRITONSERVER_REQUEST_RELEASE_ALL) != 0) {
-//    TRITONSERVER_InferenceRequestDelete(request);
-//    if (userp != nullptr) {
-//      auto warmup_promise = reinterpret_cast<std::promise<void>*>(userp);
-//      warmup_promise->set_value();
-//    }
-//  }
-//}
-
-}  // namespace
-
 Status
 InferenceBackend::GetInput(
     const std::string& name, const inference::ModelInput** input) const
@@ -273,198 +206,6 @@ InferenceBackend::Run(
   }
 
   contexts_[runner_idx]->Run(this, std::move(requests));
-}
-
-void
-InferenceBackend::WarmUp(uint32_t runner_idx, WarmupData& sample)
-{
-  // Capture timestamp before run to avoid incorrect accumulation from
-  // sequential warmup runs
-  for (auto& request : sample.requests_) {
-#ifdef TRITON_ENABLE_STATS
-    request->CaptureRequestStartNs();
-#endif  // TRITON_ENABLE_STATS
-    request->CaptureQueueStartNs();
-  }
-  Run(runner_idx, std::move(sample.requests_));
-}
-
-Status
-InferenceBackend::GenerateWarmupData(std::vector<WarmupData>* samples)
-{
-  samples->clear();
-  for (const auto& warmup_setting : config_.model_warmup()) {
-    if (warmup_setting.batch_size() == 0) {
-      LOG_VERBOSE(1) << "Skipping batch 0 warmup sample '"
-                     << warmup_setting.name() << "'";
-      continue;
-    }
-    LOG_VERBOSE(1) << "Generating warmup sample data for '"
-                   << warmup_setting.name() << "'";
-
-    // Two passes. First pass to get max byte size for synthetic
-    // data. Second pass to add original inputs and override inputs
-    // for control inputs.
-    int64_t max_zero_byte_size = 0;
-    int64_t max_random_byte_size = 0;
-    for (const auto& input_meta : warmup_setting.inputs()) {
-      auto element_count = GetElementCount(input_meta.second.dims());
-      if (element_count == -1) {
-        return Status(
-            Status::Code::INVALID_ARG,
-            "warmup setting expects all variable-size dimensions are specified "
-            "for input '" +
-                input_meta.first + "'");
-      }
-
-      int64_t batch_byte_size =
-          element_count * GetDataTypeByteSize(input_meta.second.data_type());
-      if (batch_byte_size == 0) {
-        batch_byte_size = element_count * sizeof(int32_t);
-      }
-
-      switch (input_meta.second.input_data_type_case()) {
-        case inference::ModelWarmup_Input::InputDataTypeCase::kZeroData:
-          max_zero_byte_size = std::max(batch_byte_size, max_zero_byte_size);
-          break;
-        case inference::ModelWarmup_Input::InputDataTypeCase::kRandomData: {
-          if (input_meta.second.data_type() ==
-              inference::DataType::TYPE_STRING) {
-            max_zero_byte_size = std::max(batch_byte_size, max_zero_byte_size);
-          } else {
-            max_random_byte_size =
-                std::max(batch_byte_size, max_random_byte_size);
-          }
-          break;
-        }
-        default:
-          break;
-      }
-    }
-
-    samples->emplace_back(warmup_setting.name());
-    auto& warmup_data = samples->back();
-    // Create buffers for synthetic data
-    TRITONSERVER_MemoryType type;
-    int64_t type_id;
-    warmup_data.zero_data_.reset(new AllocatedMemory(
-        max_zero_byte_size, TRITONSERVER_MEMORY_CPU_PINNED /* memory_type */,
-        0 /* memory_type_id */));
-    char* zero_buffer = warmup_data.zero_data_->MutableBuffer(&type, &type_id);
-    memset(zero_buffer, 0, max_zero_byte_size);
-
-    warmup_data.random_data_.reset(new AllocatedMemory(
-        max_random_byte_size, TRITONSERVER_MEMORY_CPU_PINNED /* memory_type */,
-        0 /* memory_type_id */));
-    char* random_buffer =
-        warmup_data.random_data_->MutableBuffer(&type, &type_id);
-    for (int64_t offset = 0; offset < max_random_byte_size; offset++) {
-      random_buffer[offset] = rand();
-    }
-
-    // Prepare the inference request for the specified sample.
-    for (size_t cnt = 0; cnt < warmup_setting.batch_size(); cnt++) {
-      warmup_data.requests_.emplace_back(new InferenceRequest(this, Version()));
-      auto& lrequest = warmup_data.requests_.back();
-
-      // Second pass to prepare original inputs.
-      std::vector<std::shared_ptr<InferenceRequest::Input>> input_sps;
-      for (const auto& input_meta : warmup_setting.inputs()) {
-        auto batch1_element_count = GetElementCount(input_meta.second.dims());
-        auto batch_byte_size =
-            batch1_element_count *
-            GetDataTypeByteSize(input_meta.second.data_type());
-        if (batch_byte_size == 0) {
-          batch_byte_size = batch1_element_count * sizeof(int32_t);
-        }
-
-        const char* allocated_ptr;
-        switch (input_meta.second.input_data_type_case()) {
-          case inference::ModelWarmup_Input::InputDataTypeCase::kZeroData:
-            allocated_ptr = zero_buffer;
-            break;
-          case inference::ModelWarmup_Input::InputDataTypeCase::kRandomData: {
-            if (input_meta.second.data_type() ==
-                inference::DataType::TYPE_STRING) {
-              allocated_ptr = zero_buffer;
-            } else {
-              allocated_ptr = random_buffer;
-            }
-            break;
-          }
-          case inference::ModelWarmup_Input::InputDataTypeCase::
-              kInputDataFile: {
-            // For data provided from file, we can set buffer in first pass
-            warmup_data.provided_data_.emplace_back(new std::string());
-            auto input_data = warmup_data.provided_data_.back().get();
-            RETURN_IF_ERROR(ReadTextFile(
-                JoinPath({model_dir_, kWarmupDataFolder,
-                          input_meta.second.input_data_file()}),
-                input_data));
-            if (input_meta.second.data_type() ==
-                inference::DataType::TYPE_STRING) {
-              batch_byte_size = input_data->size();
-            } else if (((size_t)batch_byte_size) > input_data->size()) {
-              return Status(
-                  Status::Code::INVALID_ARG,
-                  "warmup setting expects " + std::to_string(batch_byte_size) +
-                      " bytes, but the data "
-                      "provided from " +
-                      input_meta.second.input_data_file() + "only has " +
-                      std::to_string(input_data->size()) + " bytes");
-            }
-            allocated_ptr = input_data->data();
-            break;
-          }
-          default:
-            return Status(
-                Status::Code::INVALID_ARG, "warmup setting expects input '" +
-                                               input_meta.first +
-                                               "' to have input_data_type set");
-        }
-
-        const inference::ModelInput* input_config;
-        bool is_original_input =
-            GetInput(input_meta.first, &input_config).IsOk();
-        InferenceRequest::Input* input = nullptr;
-        std::vector<int64_t> input_meta_shape;
-        // Append batch size only if the model supports batching
-        // and not control inpt.
-        if ((config_.max_batch_size() != 0) && is_original_input) {
-          input_meta_shape.push_back(1);
-        }
-        for (auto d : input_meta.second.dims()) {
-          input_meta_shape.push_back(d);
-        }
-        if (is_original_input) {
-          RETURN_IF_ERROR(lrequest->AddOriginalInput(
-              input_meta.first, input_meta.second.data_type(), input_meta_shape,
-              &input));
-        } else {
-          input_sps.emplace_back();
-          RETURN_IF_ERROR(lrequest->AddOverrideInput(
-              input_meta.first, input_meta.second.data_type(),
-              (config_.max_batch_size() != 0 ? 1 : 0), input_meta_shape,
-              &input_sps.back()));
-          input = input_sps.back().get();
-        }
-        RETURN_IF_ERROR(input->AppendData(
-            allocated_ptr, batch_byte_size,
-            TRITONSERVER_MEMORY_CPU /* memory_type */, 0 /* memory_type_id */));
-      }
-
-      RETURN_IF_ERROR(lrequest->PrepareForInference());
-      // Override inputs must be added after PrepareForInference() is called
-      for (const auto& sp : input_sps) {
-        RETURN_IF_ERROR(lrequest->AddOverrideInput(sp));
-      }
-
-      RETURN_IF_ERROR(lrequest->SetResponseCallback(
-          &warmup_allocator, nullptr, WarmupResponseComplete, nullptr));
-    }
-  }
-
-  return Status::Success;
 }
 
 }}  // namespace nvidia::inferenceserver
